homework2/requirements.txt
asgiref==3.9.2
Django==5.2.7
django-bootstrap5==25.2
djangorestframework==3.16.1
sqlparse==0.5.3
